Add FromContext and Context.WithContext for lots of context
This allows building up a honeybadger.Context riding along with a
context.Context so it can be sent with errors. The goal is to make it so
that the context sent to Honeybadger is tied to the request (or stack)
rather than multiple requests clobbering the global state (see #35).

Fixes #35
Closes #37